# cloud azure: new benchmark script to get base level performance