on a local azurite server I get the following...

```
Connecting to Azure server took 0.01 seconds
performance testing...
      70.0 mb/s to write   1.0 mb to server, total time needed was 0.0 seconds
      88.9 mb/s to write   5.0 mb to server, total time needed was 0.1 seconds
     111.6 mb/s to write  10.0 mb to server, total time needed was 0.1 seconds
      89.1 mb/s to write  50.0 mb to server, total time needed was 0.6 seconds
     100.5 mb/s to write 100.0 mb to server, total time needed was 1.0 seconds
     103.6 mb/s to write 200.0 mb to server, total time needed was 1.9 seconds
```